requirements.txt
django
django-bootstrap-modal-forms
django-widget-tweaks
phonenumbers
django-phonenumber-field[phonenumbers]
Pillow
django-allauth
